feat: 공통응답구조 RsData 구현
- 정상케이스 success , 에러케이스 error 로 분리
