+++ backend/src/main/java/com/example/backend/global/rs/ErrorRs.java
@@ -1,0 +1,8 @@
+package com.example.backend.global.rs;
+
+import lombok.Builder;
+
+@Builder
+public record ErrorRs(String target, Integer code, String message) {
+
+}
